refactor: move historicData utils into recentTicketInspector package
