+++ packages/backend/api_test/GetBaseThreshold_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/FreiFahren/backend/utils"
+	"github.com/FreiFahren/backend/api/getRecentTicketInspectorInfo"
 )
 
 func TestBaseThreshold(t *testing.T) {
@@ -29,7 +29,7 @@ func TestBaseThreshold(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.dateTime, func(t *testing.T) {
 			currentTime, _ := time.Parse(layout, tc.dateTime)
-			result := utils.GetBaseThreshold(currentTime.Hour(), currentTime.Minute(), currentTime)
+			result := getRecentTicketInspectorInfo.GetBaseThreshold(currentTime.Hour(), currentTime.Minute(), currentTime)
 			roundedResult := math.Round(result*100) / 100
 			if roundedResult != tc.expected {
 				t.Errorf("BaseThreshold(%s) = %v; want %v", tc.dateTime, roundedResult, tc.expected)
